changed PayrollReportService to update amount with same pay period but with different report.

+++ payroll-service/src/main/java/com/arsal/payrollservice/dto/EmployeeReportDto.java
@@ -1,17 +1,19 @@
 package com.arsal.payrollservice.dto;
 
+import java.math.BigDecimal;
+
 public class EmployeeReportDto {
 
     public String employeeId;
 
     public PayPeriodDto payPeriodDto;
 
-    public Double amountPaid;
+    public BigDecimal amountPaid;
 
     public EmployeeReportDto() {
     }
 
-    public EmployeeReportDto(String employeeId, PayPeriodDto payPeriodDto, Double amountPaid) {
+    public EmployeeReportDto(String employeeId, PayPeriodDto payPeriodDto, BigDecimal amountPaid) {
         this.employeeId = employeeId;
         this.payPeriodDto = payPeriodDto;
         this.amountPaid = amountPaid;
@@ -33,11 +35,11 @@ public class EmployeeReportDto {
         this.payPeriodDto = payPeriodDto;
     }
 
-    public Double getAmountPaid() {
+    public BigDecimal getAmountPaid() {
         return amountPaid;
     }
 
-    public void setAmountPaid(Double amountPaid) {
+    public void setAmountPaid(BigDecimal amountPaid) {
         this.amountPaid = amountPaid;
     }
